Fixed Reset Button Bug and formatted Text-Version output
- Truncated text output values in text-based version

- Fixed error displayed in console when projectiles were removed

- Fixed Reset button bug not resetting GUI score, lives, and planet

+++ src/pdef/SpawnHandler.java
@@ -70,8 +70,8 @@ public class SpawnHandler {
 			PolarCoord p = element.getPolarCoordinates();
 			Point2D p2 = p.getRawCoordinates();
 			System.out.println("Projectile " + element.getName() 
-					+ " is " + p.getDistance() + " units away at" +
-					" (" + p2.getX() + ", " +p2.getY()+ ").");
+					+ " is " + (int)p.getDistance() + " units away at" +
+					" (" + (int)p2.getX() + ", " + (int)p2.getY()+ ").");
 		}
 	}
